+++ .gitignore
@@ -77,3 +77,24 @@ node_modules/
 !/functions/info.js
 /dist/
 !/functions/.runtimeconfig.json
+
+.DS_Store
+node_modules
+/dist
+
+
+# local env files
+.env.local
+.env.*.local
+
+# Log files
+pnpm-debug.log*
+
+# Editor directories and files
+.idea
+.vscode
+*.suo
+*.ntvs*
+*.njsproj
+*.sln
+*.sw?
