+++ package.json
@@ -1,25 +1,18 @@
 {
   "name": "atlas-project",
   "version": "0.1.0",
+  "private": true,
   "description": "The Web App for the Atlas Project.",
+  "author": {
+    "name": "Benjamin Lamont, Atlas Contributors"
+  },
   "scripts": {
     "build": "webpack --config webpack.config.js",
     "build-prod": "webpack --config prod.config.js",
-    "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook",
+    "storybook": "start-storybook -p 6006",
     "test": "firebase emulators:start"
   },
-  "private": true,
-  "repository": {
-    "type": "github",
-    "url": "https://github.com/Atlas-Project-Team/Atlas-Project.git"
-  },
-  "author": "Benjamin Lamont, Atlas Contributors",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/Atlas-Project-Team/Atlas-Project/issues"
-  },
-  "homepage": "https://github.com/Atlas-Project-Team/Atlas-Project#readme",
   "dependencies": {
     "@apollo/client": "^3.1.3",
     "apollo-server-cloud-functions": "^2.16.1",
@@ -53,6 +46,7 @@
     "@types/split": "^1.0.0",
     "@types/stats": "^0.16.30",
     "@vue/cli-plugin-typescript": "^4.5.3",
+    "@vue/cli-service": "~4.5.0",
     "babel-loader": "^8.1.0",
     "babel-preset-vue": "^2.0.2",
     "clean-webpack-plugin": "^3.0.0",
@@ -78,5 +72,15 @@
     "webpack": "^4.44.0",
     "webpack-cli": "^3.3.12",
     "webpack-node-externals": "^2.5.2"
+  },
+  "_id": "atlas-project@0.1.0",
+  "bugs": {
+    "url": "https://github.com/Atlas-Project-Team/Atlas-Project/issues"
+  },
+  "license": "GNU GPLv3",
+  "readme": "ERROR: No README data found!",
+  "repository": {
+    "type": "github",
+    "url": "git+https://github.com/Atlas-Project-Team/Atlas-Project.git"
   }
 }
