+++ src/newIndex.html
@@ -6,7 +6,7 @@
 </head>
 <body>
     <div id="root">
-        <split-buttons></split-buttons>
+
     </div>
     <script src="test.bundle.js"></script>
 </body>
