+++ stories/index.stories.js
@@ -1,4 +1,4 @@
-import SmallObjects from '../src/components/objects.vue';
+import SmallObjects from '@/components/Objects.vue';
 import client from '../src/apolloClient';
 import {gql} from "@apollo/client";
